getApplicants(projId) is moved to UserContoller and also added 'APPLIED' status to queries WHERE condition
