+++ src/main/java/org/c4sg/dao/UserDAO.java
@@ -1,8 +1,6 @@
 package org.c4sg.dao;
 
 import org.c4sg.constant.UserStatus;
-import org.c4sg.constant.UserRole;
-import org.c4sg.entity.Project;
 import org.c4sg.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
@@ -17,11 +15,11 @@ import javax.transaction.Transactional;
 
 @Component
 public interface UserDAO extends JpaRepository<User, Long>, JpaSpecificationExecutor<User> {
-
+	
     String FIND_BY_ID_QUERY = "SELECT u FROM UserProject up " +
                                 "JOIN up.user u " +
                                 "JOIN up.project p " +
-                                "WHERE p.id = :projId";
+                                "WHERE p.id =:projId and up.status= :userProjStatus";
     
     String UPDATE_SLACK_STATUS = "UPDATE User u set u.chatFlag = :isSlackReg where u.id = :userId";
     
@@ -40,5 +38,7 @@ public interface UserDAO extends JpaRepository<User, Long>, JpaSpecificationExec
     Integer updateIsSlackRegisteredFlag(@Param("isSlackReg") Boolean isSlackReg, @Param("userId") Integer userId);
   
     @Query(FIND_BY_ID_QUERY)
-    List<User> findByUserProjectId(@Param("projId") Integer projId);
+    List<User> findByUserProjectId(@Param("projId") Integer projId,
+    							   @Param("userProjStatus") String userProjStatus);
+    
 }
